Time: 85 ms (92.04%) | Memory: 91.4 MB (82.49%) - LeetSync

+++ 3332-minimum-operations-to-exceed-threshold-value-ii/minimum-operations-to-exceed-threshold-value-ii.cpp
@@ -2,15 +2,15 @@ class Solution {
 public:
     int minOperations(vector<int>& nums, int k) {
         if(nums.size()<2) return -1;
-        priority_queue<long long,vector<long long>,greater<long long>>pq(nums.begin(),nums.end());
+        priority_queue<long,vector<long>,greater<long>>pq(nums.begin(),nums.end());
         
         int count=0;
         while(pq.size()>=2 && pq.top()<k){
-            long long a=pq.top();
+            long a=pq.top();
             pq.pop();
-            long long b = pq.top();
+            long b = pq.top();
             pq.pop();
-            long long new_element= (long long)a*2+ (long long)b;
+            long new_element= a*2+ b;
             pq.push(new_element);
             count++;
         }
